+++ core-kotlin/src/main/kotlin/com/baeldung/kotlin/StringUtil.kt
@@ -1,0 +1,9 @@
+@file:JvmName("Strings")
+package com.baeldung.kotlin
+
+fun String.escapeForXml() : String {
+    return this
+            .replace("&", "&amp;")
+            .replace("<", "&lt;")
+            .replace(">", "&gt;")
+}
