+++ core-kotlin/src/test/java/com/baeldung/kotlin/StringUtilTest.java
@@ -1,0 +1,30 @@
+package com.baeldung.kotlin;
+
+import kotlin.text.StringsKt;
+import org.junit.Assert;
+import org.junit.Test;
+
+import static com.baeldung.kotlin.Strings.*;
+
+
+public class StringUtilTest {
+
+    @Test
+    public void shouldSwapTwoElementsInList() {
+        String xml = "<a>hi</a>";
+
+        String escapedXml = escapeForXml(xml);
+
+        Assert.assertEquals("&lt;a&gt;hi&lt;/a&gt;", escapedXml);
+    }
+
+    @Test
+    public void callingBuiltInKotlinExtensionMethod() {
+        String name = "john";
+
+        String capitalizedName = StringsKt.capitalize(name);
+
+        Assert.assertEquals("John", capitalizedName);
+    }
+
+}
